Changed folder structure and module exports refactoring
