+++ .gitignore
@@ -1,5 +1,4 @@
 node_modules/
 .vscode/
 .env
-logfile.log
-uncaughtExceptions.log
+logs/
