+++ controllers/article.js
@@ -1,7 +1,9 @@
 
-module.exports = {
+const articleController =  {
     getAllArticles:  async (req, res) => {
         const articles = await Article.find();
         res.send(articles);
     }
 }
+
+module.exports = articleController;
